parent 661dc4b386ab4dc1030cdadfddec692ad8d57419 author Nick Ethier <nethier@hashicorp.com> 1543376542 -0500 committer Nick Ethier <nethier@hashicorp.com> 1544122848 -0500 gpgsig -----BEGIN PGP SIGNATURE-----
 iQJKBAABCgA0FiEEYB3UisqqyVCyYZXuB8Gj7O2Q0koFAlwJceAWHG5ldGhpZXJA
 aGFzaGljb3JwLmNvbQAKCRAHwaPs7ZDSStABD/97pV2BpX+H21SbeqPnLmW/VIJg
 iTXQu3G6stAqDkIhT3swLQzaFYcvs5IXb6naXUdufQlr7J2/2KkMafVAwwpkOCOM
 bMkm6AcEMITnfKDq0M5zvZPherKqvZ/at8zbpMhXY27lgBQOL9i6i7+8jJvsmACh
 rZzeF8QuawIb1pSZM04AMY58p3dtZSu2nsurOGhg2+YCXpmJXombRL9uYo29HrcI
 496ukNO34Z+ObGeKqu+YnewolocwVY+VR/i113uEKcjzAlK4dv/HrnXW5L/ZW7cn
 03MjhNEgPfn38m6KMH+Jni53+zt0vv6OPP6iNYahlijmfe7kLVwqZmrgwwMuo8DY
 wBEfjcc7We5e1rbEjSWPCNBZAnp22UqNlR7jxeGrZJ8+CltoauEBf3Ykit1GW2w3
 JPNrnBSmIJQrKFhJzhMa+THS+5rcMSLzu3T4nUXqRV7Zi32Lo0iQmBKI7ioNhYo0
 z8bINN5igxiGoOwg0g9rwyZf7P51EeygwkYzA9ggG9n1tDreS8QCM6U/NQmQE3dz
 aGGg9vnm59GmhHC3Ij2hrEjToQlqNq7m2F7gsyvlCbsh27YvZmNiFOZM6NQjVxRR
 EZttpLL/YgBEHhzl5pd076TGfUIdMlPSbdprz8WjbErwwhO6Gy4i0Ke2wh1hMm6H
 Ic3TykxcjO0mQlfvjw==
 =cJ78
 -----END PGP SIGNATURE-----

client/drivermananger: add driver manager

The driver manager is modeled after the device manager and is started by the client.
It's responsible for handling driver lifecycle and reattachment state, as well as
processing the incomming fingerprint and task events from each driver. The mananger
exposes a method for registering event handlers for task events that is used by the
task runner to update the server when a task has been updated with an event.

Since driver fingerprinting has been implemented by the driver manager, it is no
longer needed in the fingerprint mananger and has been removed.